Update by Merge for bitdubai

+++ BNK/library/api/fermat-bnk-api/src/main/java/com/bitdubai/fermat_bnk_api/all_definition/bank_money_transaction/BankTransactionParameters.java
@@ -2,6 +2,7 @@ package com.bitdubai.fermat_bnk_api.all_definition.bank_money_transaction;
 
 import com.bitdubai.fermat_api.layer.all_definition.enums.FiatCurrency;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 /**
@@ -19,7 +20,7 @@ public interface BankTransactionParameters {
     String getPublicKeyActor();
 
 
-    float getAmount();
+    BigDecimal getAmount();
 
 
     String getAccount();
